extend the server pool shutdown test case
which now features a test for re-run protection and a test for
the server basepath's removal throwing an OSError

+++ webradio/pool.py
@@ -44,8 +44,13 @@ class Server(object):
             yield worker.socket
 
     def shutdown(self):
+        # don't do anything if we already shut down
+        if not self.workers:
+            return
+
         for worker in self.workers:
             worker.shutdown()
+        self.workers = []
 
         with ignore(OSError):
             self.basepath.rmdir()
